Create structureProblem.c
Write down a structure capable of storing the following information about a student
registered in a course:
❖ ID
❖ Gender
❖ Class test score
❖ Mid-term score
❖ Final Exam score
❖ Total score
(ii) Use the above structure to take input of N students’ information as input in an
array of structure. N will also be input to your program.
(iii) In the same program, write down a function that will take students’ information
(within an array of structure) and the number of the students as parameters and returns
how many of them failed. Assume that a student needs to obtain total 40 in order to pass
in the exam.
(iv) In the same program, write down another function that will take students’
information (within an array of structure) and the number of students as parameters and
returns the ID of the student obtaining the highest marks.